Move pulsar plugin to optional plugins
