+++ apm-protocol/apm-network/src/main/java/org/apache/skywalking/apm/network/trace/component/ComponentsDefine.java
@@ -136,8 +136,6 @@ public class ComponentsDefine {
   
     public static final OfficialComponent LIGHT_4J = new OfficialComponent(71, "Light4J");
 
-    public static final OfficialComponent PULSAR = new OfficialComponent(72, "Pulsar");
-
     public static final OfficialComponent PULSAR_PRODUCER = new OfficialComponent(73, "pulsar-producer");
 
     public static final OfficialComponent PULSAR_CONSUMER = new OfficialComponent(74, "pulsar-consumer");
