+++ apm-sniffer/optional-plugins/pulsar-plugin/pom.xml
@@ -19,7 +19,7 @@
 
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>apm-sdk-plugin</artifactId>
+        <artifactId>optional-plugins</artifactId>
         <groupId>org.apache.skywalking</groupId>
         <version>6.5.0-SNAPSHOT</version>
     </parent>
