+++ apm-sniffer/optional-plugins/pulsar-plugin/src/main/java/org/apache/skywalking/apm/plugin/pulsar/PulsarConsumerInterceptor.java
@@ -32,6 +32,23 @@ import org.apache.skywalking.apm.network.trace.component.ComponentsDefine;
 
 import java.lang.reflect.Method;
 
+/**
+ * Interceptor for pulsar consumer enhanced instance
+ *
+ * Here is the intercept process steps:
+ *
+ * <pre>
+ *  1. Get the @{@link ConsumerEnhanceRequiredInfo} and record the service url, topic name and subscription name
+ *  2. Create the entry span when call <code>messageProcessed</code> method
+ *  3. Extract all the <code>Trace Context</code> when call <code>messageProcessed</code> method
+ *  4. Stop the entry span when <code>messageProcessed</code> method finished.
+ * </pre>
+ *
+ * If the processed message is null or arguments length is 0 will record exception for the active span
+ * and notify the active span error occurred.
+ *
+ * @author penghui
+ */
 public class PulsarConsumerInterceptor implements InstanceMethodsAroundInterceptor {
 
     public static final String OPERATE_NAME_PREFIX = "Pulsar/";
@@ -40,43 +57,41 @@ public class PulsarConsumerInterceptor implements InstanceMethodsAroundIntercept
     @Override
     public void beforeMethod(EnhancedInstance objInst, Method method, Object[] allArguments, Class<?>[] argumentsTypes,
                              MethodInterceptResult result) throws Throwable {
-        ConsumerEnhanceRequiredInfo requiredInfo = (ConsumerEnhanceRequiredInfo) objInst.getSkyWalkingDynamicField();
-        requiredInfo.setStartTime(System.currentTimeMillis());
-        AbstractSpan activeSpan = ContextManager.createEntrySpan(OPERATE_NAME_PREFIX +
-                requiredInfo.getTopic() + CONSUMER_OPERATE_NAME + requiredInfo.getSubscriptionName(), null)
-                .start(requiredInfo.getStartTime());
-        activeSpan.setComponent(ComponentsDefine.PULSAR_CONSUMER);
-        SpanLayer.asMQ(activeSpan);
-        Tags.MQ_BROKER.set(activeSpan, requiredInfo.getServiceUrl());
-        Tags.MQ_TOPIC.set(activeSpan, requiredInfo.getTopic());
+        if (allArguments.length > 0 && allArguments[0] != null) {
+            ConsumerEnhanceRequiredInfo requiredInfo = (ConsumerEnhanceRequiredInfo) objInst.getSkyWalkingDynamicField();
+            requiredInfo.setStartTime(System.currentTimeMillis());
+            AbstractSpan activeSpan = ContextManager.createEntrySpan(OPERATE_NAME_PREFIX +
+                    requiredInfo.getTopic() + CONSUMER_OPERATE_NAME + requiredInfo.getSubscriptionName(), null)
+                    .start(requiredInfo.getStartTime());
+            activeSpan.setComponent(ComponentsDefine.PULSAR_CONSUMER);
+            SpanLayer.asMQ(activeSpan);
+            Tags.MQ_BROKER.set(activeSpan, requiredInfo.getServiceUrl());
+            Tags.MQ_TOPIC.set(activeSpan, requiredInfo.getTopic());
+            Message msg = (Message) allArguments[0];
+            ContextCarrier contextCarrier = new ContextCarrier();
+            CarrierItem next = contextCarrier.items();
+            while (next.hasNext()) {
+                next = next.next();
+                next.setHeadValue(msg.getProperty(next.getHeadKey()));
+            }
+            ContextManager.extract(contextCarrier);
+        }
     }
 
     @Override
     public Object afterMethod(EnhancedInstance objInst, Method method, Object[] allArguments, Class<?>[] argumentsTypes,
                               Object ret) throws Throwable {
-        if (allArguments.length > 0) {
-            Message msg = (Message) allArguments[0];
-            if (null != msg) {
-                ContextCarrier contextCarrier = new ContextCarrier();
-                CarrierItem next = contextCarrier.items();
-                while (next.hasNext()) {
-                    next = next.next();
-                    next.setHeadValue(msg.getProperty(next.getHeadKey()));
-                }
-                ContextManager.extract(contextCarrier);
-            } else {
-                ContextManager.activeSpan().errorOccurred().log(new RuntimeException("Processed messages is null!"));
-            }
-        } else {
-            ContextManager.activeSpan().errorOccurred().log(new RuntimeException("Call messageProcessed() with 0 arguments!"));
+        if (allArguments.length > 0 && allArguments[0] != null) {
+            ContextManager.stopSpan();
         }
-        ContextManager.stopSpan();
         return ret;
     }
 
     @Override
     public void handleMethodException(EnhancedInstance objInst, Method method, Object[] allArguments,
                                       Class<?>[] argumentsTypes, Throwable t) {
-        ContextManager.activeSpan().errorOccurred().log(t);
+        if (allArguments.length > 0 && allArguments[0] != null) {
+            ContextManager.activeSpan().errorOccurred().log(t);
+        }
     }
 }
