+++ apm-sniffer/optional-plugins/pulsar-plugin/src/main/resources/skywalking-plugin.def
